feat (#202) :: 실행되지 않는 알람 예외 적용

+++ src/main/java/team/jeonghokim/daedongyeojido/domain/user/service/DeleteUserAlarmService.java
@@ -4,9 +4,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team.jeonghokim.daedongyeojido.domain.alarm.domain.UserAlarm;
+import team.jeonghokim.daedongyeojido.domain.alarm.domain.enums.AlarmCategory;
 import team.jeonghokim.daedongyeojido.domain.alarm.domain.repository.UserAlarmRepository;
 import team.jeonghokim.daedongyeojido.domain.alarm.exception.AlarmAccessDeniedException;
 import team.jeonghokim.daedongyeojido.domain.alarm.exception.AlarmNotFoundException;
+import team.jeonghokim.daedongyeojido.domain.alarm.exception.CannotDeleteAlarmException;
 import team.jeonghokim.daedongyeojido.domain.user.domain.User;
 import team.jeonghokim.daedongyeojido.domain.user.facade.UserFacade;
 
@@ -27,6 +29,10 @@ public class DeleteUserAlarmService {
 
         if (!user.getId().equals(alarm.getReceiver().getId())) {
             throw AlarmAccessDeniedException.EXCEPTION;
+        }
+
+        if (!alarm.getAlarmCategory().equals(AlarmCategory.COMMON) && !alarm.isExecuted()) {
+            throw CannotDeleteAlarmException.EXCEPTION;
         }
 
         userAlarmRepository.delete(alarm);
